+++ modules/transport-netty4/src/main/java/org/opensearch/transport/Netty4Plugin.java
@@ -32,8 +32,8 @@
 
 package org.opensearch.transport;
 
-import org.apache.lucene.util.SetOnce;
 import org.opensearch.Version;
+import org.opensearch.common.SetOnce;
 import org.opensearch.common.io.stream.NamedWriteableRegistry;
 import org.opensearch.common.network.NetworkModule;
 import org.opensearch.common.network.NetworkService;